playerが0のルームを削除
